requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
openai==0.27.8
python-dotenv==1.0.0
